mod: change extension from .js to .ts
